requirements.txt
dotmotif==0.13.0
fastapi>=0.88.0
neuprint_python>=0.4.25
networkx>=2.8.8
requests>=2.28.1
tamarind>=0.2.0
gunicorn==20.0.4
uvicorn[standard]>=0.20.0